Move helper funcs to utils
