+++ tests/test_cov_entry.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-import analyze
+import utils
 from fitting import FitResult
 
 
@@ -24,15 +24,15 @@ def test_cov_entry_valid_params():
         ]
     )
     fr = FitResult(params, cov, 0)
-    assert analyze._cov_entry(fr, "A", "B") == pytest.approx(0.1)
-    assert analyze._cov_entry(fr, "C", "A") == pytest.approx(0.2)
-    assert analyze._cov_entry(fr, "B", "C") == pytest.approx(0.3)
+    assert utils._cov_entry(fr, "A", "B") == pytest.approx(0.1)
+    assert utils._cov_entry(fr, "C", "A") == pytest.approx(0.2)
+    assert utils._cov_entry(fr, "B", "C") == pytest.approx(0.3)
 
 
 def test_cov_entry_missing_or_none():
     params = {"A": 1.0, "B": 2.0}
     cov = np.eye(2)
     fr = FitResult(params, cov, 0)
-    assert analyze._cov_entry(fr, "A", "C") == 0.0
+    assert utils._cov_entry(fr, "A", "C") == 0.0
     fr_none = FitResult(params, None, 0)
-    assert analyze._cov_entry(fr_none, "A", "B") == 0.0
+    assert utils._cov_entry(fr_none, "A", "B") == 0.0
